Sensor server working, but is having trouble with UI
The idea for for updating the sensor UI is there but I have to comment
out the creation of the UI workers(both time and sensor update) because
they break the program.

I(bill) feel that there is some sort of race condition in the current UI
server/worker design. I was trying to follow the update time but neither
of which I got working. Perhaps we need some sort of syncrounization
between server and worker.