Various bug fixes/robustness changes to track Manager
including:
1. it was ambigous where the position E12 C6 500 is on track A because
there are two ways of going from E12 to C6, one passing two sensors and
one passing two switches. I am tightening what we can use as Positions
to consecutive sensors so we know it represents the path passing two
switches.
2. Fixed a bug in traversing/locating a position where D6 E10 X was being
treated as D6 D9 X,
3. Increase robustness on trackManger handling position when offest is
negative or bigger than the distance between the two landmarks. Now
track treats the position as on top of landmark1 for the first case and
on top of landmkark2 for the second case.